Using the Node Modules System
removed all the routing logic inside the app.js and put it in the newly created file called route.js
the routing logic is put inside inside an arrow function called routeHandler
exported the routeHandler function from routes.js and import it in app.js
used the exported function inside http.createServer

closes #11
